feat(actions): migrate share link management to Astro Actions
- Add share.create and share.revoke actions in src/actions/index.ts
- Update VideoHeader.tsx to call actions instead of fetch
- Remove unused /api/share/manage/[videoId].ts route

Closes #60
